app/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-JWT-Extended==3.17.0
mysqlclient==1.3.13
Werkzeug==0.14.1
Flask-Cors==3.0.7
requests-oauthlib==1.2.0
boto3==1.9.115
uWSGI==2.0.18
Flask-Migrate==2.4.0
coverage==4.5.3
Flask-Testing==0.7.1
nose==1.3.7